之前未追踪文件，以及解决修改订单id类型为long ，删除订单报错

+++ src/main/java/com/hbnu/controller/OrdersInfoController.java
@@ -178,7 +178,7 @@ public class OrdersInfoController {
 
     @RequestMapping("/deloneorder")
     @ResponseBody
-    public Result delOneOrder(@RequestParam("id")int id){
+    public Result delOneOrder(@RequestParam("id")long id){
         ius.deleteOneOrder(id);
         Result result = Result.success();
         return result;
